backend/requirements_ml.txt
# FastAPI ML Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
tensorflow==2.15.0
python-dotenv==1.0.0

# Optional: For GPU support
# tensorflow-gpu==2.15.0
